refactor: Move out dynamic icons
